banyabot/requirements.txt
langdetect==1.0.9
mysql-connector-python==8.0.33
openai==0.27.0
pyTelegramBotAPI==4.7.0
python-telegram-bot==20.3

